Debugged sending config via msg, seems good, even takes new config while running

+++ joystickRC_structs.h
@@ -3,15 +3,15 @@
 struct messageState {
 
 	unsigned char *messageBuffer;
-	byte readBytes;
-	byte length;
+	int readBytes;
+	int length;
 
 };
 
 struct ledBlinker {
   
 	unsigned int interval; // interval == 0 is always on, interval < 0 is always off, interval > 0 is an interval
-	byte pin;
+	int pin;
 	boolean state;
 	unsigned long lastChanged;
   
@@ -19,23 +19,23 @@ struct ledBlinker {
 
 struct configValues {
 
-	byte debugPin;                // Pin for debug LED
-	byte statusLEDPin;            // Pin for status LED
-	byte navlightPin;             // Pin for navlight LED/LEDs
-	byte rssiPin;                 // RSSI input pin
-	byte mainBatteryPin;          // Main battery input pin
-	byte commBatteryPin;          // Comm battery input pin (this board)
-	byte videoBatteryPin;         // Video battery input pin
+	int debugPin;                // Pin for debug LED
+	int statusLEDPin;            // Pin for status LED
+	int navlightPin;             // Pin for navlight LED/LEDs
+	int rssiPin;                 // RSSI input pin
+	int mainBatteryPin;          // Main battery input pin
+	int commBatteryPin;          // Comm battery input pin (this board)
+	int videoBatteryPin;         // Video battery input pin
 	int lostMessageThreshold;     // Time in ms without a message before assuming we've lost our signal
 	int heartbeatInterval;        // Interval in ms to send our heartbeat
 	int pingInterval;             // Interval in ms to send pings
-	byte servoCount;              // # of servos on this board
-	byte buttonCount;             // # of buttons on this board
-	byte *buttonPinMap;	      // Mapping of buttons to output pins
+	int servoCount;              // # of servos on this board
+	int buttonCount;             // # of buttons on this board
+	int *buttonPinMap;	      // Mapping of buttons to output pins
 	int ppmMinPulse;              // PPM min pulse length in 1/2 usec (default = 2000)
 	int ppmMaxPulse;              // PPM max pulse length in 1/2 usec (default = 4000)
 	int ppmHighPulse;             // PPM high pulse duration in 1/2 usec (default = 400)
-	byte ppmPulses;               // How many pulses are there (servo_count * 2) + 2
+	int ppmPulses;               // How many pulses are there (servo_count * 2) + 2
 	int ppmSyncPulse;             // PPM sync pulse length (bunch of math determines this)
 	int statusIntervalSignalLost; // Interval for status to flash when signal is lost
 	int statusIntervalOK;         // Interval for status to flash when everything is OK
@@ -50,13 +50,13 @@ struct signalState {
 	unsigned char pingData;
 	unsigned long lastMessageTime; // Time of last legit message, -100 initially so the PPM won't turn on until we get a real message
 	unsigned long lastMessageSentTime;
-	byte ctrlCounter;
+	int ctrlCounter;
 
 };
 
 struct ppmState {
 
-	byte currentPulse; // The pulse being sent
+	int currentPulse; // The pulse being sent
 	boolean ppmON;     // Is PPM on
 	int *pulses;       // How many PPM pulses are there
